Add a message type to clear frontend messages
This will be sent by backend prior to loading messages of a
session after an init.

Even though we clear this on the frontend, the backend might
still have messages in the pipe being sent over that needs to be
cleared.

This can be further improved if we include session id to the
messages so we can do some filtering on the frontend for
messages that aren't relevant to the session.